Added coloring to the 10 boxes
Thought this could be a potential use case for
!important. This way the background color
of black could be left on the original class
declaration for the boxes. That way if new
boxes are added they would have a default
color of black. And just wanted to try it out
since it’s never really used.